refactor: abstract application entry point to a propper class
